refactor interfaces into their own files
