+++ src/Item.ts
@@ -1,11 +1,12 @@
 import { Response } from ".";
 import { Constants } from "./common";
 import { Container } from "./Container";
-import { DocumentClient, RequestOptions } from "./documentclient";
+import { CosmosClient } from "./CosmosClient";
+import { RequestOptions } from "./request/RequestOptions";
 
 export class Item {
 
-    private client: DocumentClient;
+    private client: CosmosClient;
     public get url() {
         return `${this.container.url}/${Constants.Path.DocumentsPathSegment}/${this.id}`;
     }
@@ -14,27 +15,27 @@ export class Item {
         public readonly container: Container,
         public readonly id: string,
         public readonly primaryKey: string) {
-        this.client = this.container.database.client.documentClient;
+        this.client = this.container.database.client;
     }
 
     public read<T>(options?: RequestOptions): Promise<Response<T>> {
         if (!options.partitionKey && this.primaryKey) {
             options.partitionKey = this.primaryKey;
         }
-        return this.client.readDocument(this.url, options) as Promise<Response<T>>;
+        return this.client.documentClient.readDocument(this.url, options) as Promise<Response<T>>;
     }
 
     public replace<T>(body: T, options?: RequestOptions): Promise<Response<T>> {
         if (!options.partitionKey && this.primaryKey) {
             options.partitionKey = this.primaryKey;
         }
-        return this.client.replaceDocument(this.url, body, options) as Promise<Response<T>>;
+        return this.client.documentClient.replaceDocument(this.url, body, options) as Promise<Response<T>>;
     }
 
     public delete<T>(options?: RequestOptions): Promise<Response<T>> {
         if (!options.partitionKey && this.primaryKey) {
             options.partitionKey = this.primaryKey;
         }
-        return this.client.deleteDocument(this.url, options) as Promise<Response<T>>;
+        return this.client.documentClient.deleteDocument(this.url, options) as Promise<Response<T>>;
     }
 }
