+++ src/Offer.ts
@@ -1,8 +1,8 @@
 import { Response } from ".";
 import { Constants } from "./common";
 import { CosmosClient } from "./CosmosClient";
-import { RequestOptions } from "./documentclient";
 import { OfferDefinition } from "./OfferDefinition";
+import { RequestOptions } from "./request/RequestOptions";
 
 export class Offer {
