+++ src/Trigger.ts
@@ -2,8 +2,8 @@ import { Response } from ".";
 import { Constants } from "./common";
 import { Container } from "./Container";
 import { CosmosClient } from "./CosmosClient";
-import { RequestOptions } from "./documentclient";
 import { TriggerDefinition } from "./TriggerDefinition";
+import { RequestOptions } from "./request/RequestOptions";
 
 export class Trigger {
     public get url() {
