+++ src/Triggers.ts
@@ -1,9 +1,10 @@
 import { Container } from "./Container";
 import { CosmosClient } from "./CosmosClient";
-import { FeedOptions, RequestOptions } from "./documentclient";
 import { SqlQuerySpec } from "./queryExecutionContext";
 import { QueryIterator } from "./queryIterator";
-import { Response } from "./request";
+import { FeedOptions } from "./request/FeedOptions";
+import { Response } from "./request/request";
+import { RequestOptions } from "./request/RequestOptions";
 import { Trigger } from "./Trigger";
 import { TriggerDefinition } from "./TriggerDefinition";
