+++ src/index.ts
@@ -7,7 +7,7 @@ export { HashPartitionResolver } from "./hash";
 export { Constants, UriFactory } from "./common";
 export { Base } from "./base";
 export { RetryOptions } from "./retry";
-export { Response } from "./request";
+export { Response } from "./request/request";
 
 export { IHeaders } from "./queryExecutionContext";
 export { QueryIterator } from "./queryIterator";
