+++ src/queryExecutionContext/EndpointComponent/TopEndpointComponent.ts
@@ -1,5 +1,4 @@
-import { IHeaders } from "..";
-import { Response } from "../../request";
+import { Response } from "../../request/request";
 import { IExecutionContext } from "../IExecutionContext";
 import { IEndpointComponent } from "./IEndpointComponent";
