+++ src/queryExecutionContext/IHeaders.ts
@@ -1,0 +1,3 @@
+export interface IHeaders {
+    [key: string]: string | boolean | number;
+}
