+++ src/request/Response.ts
@@ -1,0 +1,6 @@
+import { IHeaders } from "..";
+
+export interface Response<T> {
+    headers?: IHeaders;
+    result?: T;
+}
